requirements.txt
geopandas
streamlit_folium
folium
pandas
shapely

mapclassify
matplotlib